[R7D6] 复原因公告导致的一些问题
复原导航栏的展开功能
公告支持刷新功能，并显示更新时间

+++ src/modules/subWidgetFunctions/announcementFunctions.py
@@ -1,4 +1,5 @@
 import os
+import re
 import sys
 
 from PyQt6.QtGui import QIcon, QPixmap
@@ -6,33 +7,19 @@ from PyQt6.QtWidgets import QListWidgetItem
 
 sys.path.append("..")
 
-import json
 import requests
 from components import infoBars, downloader
 from components.OSUtils import getWorkingDir, HTML_MODEL
 
 
-def contentHTMLPhaser(contentHTML):
-    contentHTML = contentHTML.replace("0.1rem", "1rem")
-    contentHTML = contentHTML.replace("0.10rem", "1rem")
-    contentHTML = contentHTML.replace("0.12rem", "1rem")
-    contentHTML = contentHTML.replace("line-height: 2", "line-height: 1.5")
-    contentHTML = contentHTML.replace("min-height: 1.5em;", "min-height: 1em;")
-    contentHTML = contentHTML.replace("white-space: pre-wrap;", "")
+def contentHTMLPhaser(css, contentHTML):
+    contentHTML = contentHTML.replace('''style="color:rgba(85,85,85,1)"''', "")
+    contentHTML = contentHTML.replace('''style="background-color: rgb(255, 215, 185);"''', "")
+    contentHTML = contentHTML.replace('''style="background-color: rgb(254, 245, 231);"''', "")
     contentHTML = contentHTML.replace('''&lt;t class="t_lc"&gt;''', "")
     contentHTML = contentHTML.replace('''&lt;t class="t_gl"&gt;''', "")
     contentHTML = contentHTML.replace('''&lt;/t&gt;''', "")
-    contentHTML = contentHTML.replace('''javascript:miHoYoGameJSSDK.openInWebview(\'''', "")
-    contentHTML = contentHTML.replace("')", "")
-    contentHTML = contentHTML.replace('''<p style="">''', '''<p style="color: white;">''')
-    contentHTML = contentHTML.replace("href=\"h", '')
-    contentHTML = contentHTML.replace("<span ", "<p ").replace("</span>", "</p>")
-    contentHTML = contentHTML.replace("<strong>", "").replace("<,strong>", "")
-    contentHTML = contentHTML.replace("rgba(51,51,51,1)", "rgba(220,220,220,1)")
-    contentHTML = contentHTML.replace("rgba(65,70,75,1)", "rgba(220,220,220,1)")
-    contentHTML = contentHTML.replace("rgba(85,85,85,1)", "rgba(220,220,220,1)")
-    contentHTML = contentHTML.replace("rgba(236,73,35,1)", "rgba(220,220,220,1)")
-    return HTML_MODEL.replace("{content}", contentHTML)
+    return HTML_MODEL.replace("{css}", css).replace("{content}", contentHTML)
 
 
 class AnnouncementFunctions:
@@ -44,17 +31,16 @@ class AnnouncementFunctions:
         self.announceIconNameList = []
         self.announceTitleList = []
 
+        self.announceStyle = open(f"{getWorkingDir()}/assets/css/github-markdown-dark.css", 'r').read()
+
         self.getIconsURL()
         self.getTitles()
 
     def getIconsURL(self):
-        try:
-            for first_level in self.announceIconData:
-                for second_level in first_level["list"]:
-                    self.announceIconURLList.append(second_level["tag_icon"])
-                    self.announceIconNameList.append(second_level["tag_icon"].split("/")[-1])
-        except requests.exceptions.ConnectionError:
-            return "ConnectionError"
+        for first_level in self.announceIconData:
+            for second_level in first_level["list"]:
+                self.announceIconURLList.append(second_level["tag_icon"])
+                self.announceIconNameList.append(second_level["tag_icon"].split("/")[-1])
 
     def getIcons(self):
         for eachIconURL in self.announceIconURLList:
@@ -74,24 +60,33 @@ class AnnouncementFunctions:
                                                 self.announceTitleList[eachItem]))
         return sideBarItems
 
+    def getImageURLFromSource(self, source):
+        pattern = re.compile(r'https://webstatic.mihoyo.com/upload/ann/[^\s]+.jpg')
+        url_lst = pattern.findall(source)
+        return url_lst
+
     def getCurrentAnnounce(self, index):
         currentAnnounce = self.announceData["list"][index]
         announceId = str(currentAnnounce["ann_id"])
         bigTitle = currentAnnounce["title"]
         contentHtml = currentAnnounce["content"]
-        if not os.path.exists(f"{getWorkingDir()}/cache/{announceId}.jpg"):
-            downloader.downloadFromImage(currentAnnounce["banner"], f"{getWorkingDir()}/cache/", announceId + ".jpg")
-        banner = QPixmap(f"{getWorkingDir()}/cache/{announceId}.jpg")
-        bannerSize = banner.rect().getRect()
-        if banner.rect().getRect() == (0, 0, 0, 0):
-            os.remove(f"{getWorkingDir()}/cache/{announceId}.jpg")
-            try:
-                downloader.downloadFromImage(currentAnnounce["banner"], f"{getWorkingDir()}/cache/",
-                                             announceId + ".jpg")
-            except requests.exceptions.MissingSchema:
-                pass
+        if currentAnnounce["banner"]:
+            downloader.downloadFromImage(currentAnnounce["banner"], f"{getWorkingDir()}/cache/", announceId + ".jpg") if not os.path.exists(f"{getWorkingDir()}/cache/{announceId}.jpg") else None
             banner = QPixmap(f"{getWorkingDir()}/cache/{announceId}.jpg")
-        bannerHeight = round(750 / bannerSize[2] * bannerSize[3])
-        contentHtml = contentHTMLPhaser(contentHtml)
+            bannerSize = banner.rect().getRect()
+            if banner.rect().getRect() == (0, 0, 0, 0):
+                os.remove(f"{getWorkingDir()}/cache/{announceId}.jpg")
+                try:
+                    downloader.downloadFromImage(currentAnnounce["banner"], f"{getWorkingDir()}/cache/",
+                                                 announceId + ".jpg")
+                except requests.exceptions.MissingSchema:
+                    pass
+                banner = QPixmap(f"{getWorkingDir()}/cache/{announceId}.jpg")
+            bannerHeight = round(750 / bannerSize[2] * bannerSize[3])
+        else:
+            banner = ""
+            bannerHeight = 0
+
+        contentHtml = contentHTMLPhaser(self.announceStyle, contentHtml)
         return {"announceId": announceId, "bigTitle": bigTitle, "banner": banner, "bannerHeight": bannerHeight,
                 "contentHtml": contentHtml}
